refactore code and using test category
